Change `ColorId` to be `UInt16`.

+++ gd/schema/aliases/id.capnp
@@ -5,7 +5,7 @@ using import "../option.capnp".OptionUInt64;
 using Id = UInt64;
 using OptionId = OptionUInt64;
 
-using ColorId = UInt8;
+using ColorId = UInt16;
 
 using IconId = UInt8;
 using LongIconId = UInt16;
